--- integrations/cypress/cypress/support/commands/authorization_code/logout.ts
@@ -1,12 +1,0 @@
-import { CommandMaker } from '../../types';
-import { makeCypressLogger } from '../../utils/cypress-logger';
-
-const log = makeCypressLogger('simulacrum-logout-ac');
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const makeAuthorizationCodeLogout = ({ atom, getClientFromSpec }: CommandMaker) => () => {
-  log('logging out');
-  return cy.request('/api/auth/logout')
-            .destroySimulation()
-            .reload();
-};
