--- integrations/cypress/cypress/support/commands/authorization_code_with_pkce/logout.ts
@@ -1,13 +1,0 @@
-import { makeCypressLogger } from '../../utils/cypress-logger';
-
-const log = makeCypressLogger('simulacrum-logout-pkce');
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const makeLogoutWithPKCE = () => () => {
-  log('logging out');
-
-  return cy
-    .destroySimulation()
-    .clearCookies()
-    .reload();
-};
